I simplified the code, made a working error-message and an endpoint that returns a single result.
